backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=1.10.0,<3.0.0
ollama==0.3.2
requests>=2.28.0
pymongo>=4.5.0
motor>=3.3.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
